Changed quizzes to render 1 per difficulty
added template for submitScore, updated styling
