+++ frontend/src/components/quiz/quiz.css
@@ -4,6 +4,7 @@
     text-align: center;
     color: greenyellow;
     padding: 5%;
+    font-size: 2rem;
 }
 
 #scoreCounter {
